scripts/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
langdetect>=1.0.9
deep-translator>=1.11.4
pillow>=10.0.0
lxml>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.1
